refactor: separate api and controller QUZ-74
