+++ game-service/game-application/app-api/src/main/kotlin/com/grepp/quizy/game/api/game/GameChatController.kt
@@ -5,13 +5,13 @@ import com.grepp.quizy.game.domain.game.GameChatService
 import org.springframework.messaging.handler.annotation.DestinationVariable
 import org.springframework.messaging.handler.annotation.MessageMapping
 import org.springframework.messaging.handler.annotation.Payload
+import org.springframework.stereotype.Controller
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 import java.security.Principal
 
-@RestController
-@RequestMapping("/api/game/chat")
-class GameChatApi(
+@Controller
+class GameChatController(
     private val gameChatService: GameChatService
 ) {
 
@@ -27,4 +27,5 @@ class GameChatApi(
             request.message
         )
     }
+
 }
