+++ game-service/game-application/app-api/src/main/kotlin/com/grepp/quizy/game/api/game/GameMatchingApi.kt
@@ -31,15 +31,4 @@ class GameMatchingApi(
         )
     )
 
-    @MessageMapping("/join/{gameId}")
-    fun joinRandomGame(
-        @DestinationVariable gameId: Long,
-        principal: Principal
-    ) {
-        gameMatchingService.join(
-            principal.name.toLong(),
-            gameId
-        )
-    }
-
 }
