+++ game-service/game-application/app-api/src/main/kotlin/com/grepp/quizy/game/api/game/GameMatchingController.kt
@@ -1,0 +1,25 @@
+package com.grepp.quizy.game.api.game
+
+import com.grepp.quizy.game.domain.game.GameMatchingService
+import org.springframework.messaging.handler.annotation.DestinationVariable
+import org.springframework.messaging.handler.annotation.MessageMapping
+import org.springframework.stereotype.Controller
+import java.security.Principal
+
+@Controller
+class GameMatchingController(
+    private val gameMatchingService: GameMatchingService
+) {
+
+    @MessageMapping("/join/{gameId}")
+    fun joinRandomGame(
+        @DestinationVariable gameId: Long,
+        principal: Principal
+    ) {
+        gameMatchingService.join(
+            principal.name.toLong(),
+            gameId
+        )
+    }
+
+}
