+++ game-service/game-application/app-api/src/main/kotlin/com/grepp/quizy/game/api/game/GamePlayingController.kt
@@ -5,11 +5,12 @@ import com.grepp.quizy.game.domain.game.GamePlayService
 import org.springframework.messaging.handler.annotation.DestinationVariable
 import org.springframework.messaging.handler.annotation.MessageMapping
 import org.springframework.messaging.handler.annotation.Payload
+import org.springframework.stereotype.Controller
 import org.springframework.web.bind.annotation.RestController
 import java.security.Principal
 
-@RestController
-class GamePlayingApi(
+@Controller
+class GamePlayingController(
     private val gamePlayService: GamePlayService
 ) {
