+++ game-service/game-application/app-api/src/main/kotlin/com/grepp/quizy/game/api/game/GamePrivateApi.kt
@@ -45,68 +45,4 @@ class GamePrivateApi(
             )
         )
 
-    @MessageMapping("/quit/{gameId}")
-    fun quit(
-        @DestinationVariable gameId: Long,
-        principal: Principal
-    ): ApiResponse<Unit> =
-        ApiResponse.success(
-            gamePrivateService.quit(
-                principal.name.toLong(),
-                gameId
-            )
-        )
-
-    @MessageMapping("/update/{gameId}/subject")
-    fun updateSubject(
-        @DestinationVariable gameId: Long,
-        @Payload request: UpdateSubjectPayloadRequest,
-        principal: Principal
-    ) {
-        gamePrivateService.updateSubject(
-            principal.name.toLong(),
-            gameId,
-            request.subject
-        )
-    }
-
-    @MessageMapping("/update/{gameId}/level")
-    fun updateLevel(
-        @DestinationVariable gameId: Long,
-        @Payload request: UpdateLevelPayloadRequest,
-        principal: Principal
-    ) {
-        gamePrivateService.updateLevel(
-            principal.name.toLong(),
-            gameId,
-            request.level
-        )
-    }
-
-    @MessageMapping("/update/{gameId}/quiz-count")
-    fun updateQuizCount(
-        @DestinationVariable gameId: Long,
-        @Payload request: UpdateQuizCountPayloadRequest,
-        principal: Principal
-    ) {
-        gamePrivateService.updateQuizCount(
-            principal.name.toLong(),
-            gameId,
-            request.quizCount
-        )
-    }
-
-    @MessageMapping("/kick/{gameId}")
-    fun kickUser(
-        @DestinationVariable gameId: Long,
-        @Payload request: KickUserPayloadRequest,
-        principal: Principal
-    ) {
-        gamePrivateService.kickUser(
-            principal.name.toLong(),
-            gameId,
-            request.targetUserId
-        )
-    }
-
 }
